LPS-153267 Simplify using a viewsReducer file
