--- modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/actions/updateViewComponent.js
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-export const ACTION_UPDATE_VIEW_COMPONENT = 'ACTION_UPDATE_VIEW_COMPONENT';
-
-export function updateViewComponent(name, component) {
-	return {
-		type: ACTION_UPDATE_VIEW_COMPONENT,
-		value: {component, name},
-	};
-}
