+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/thunks/persistVisibleFieldNames.js
@@ -12,8 +12,8 @@
  * details.
  */
 
-import {updateVisibleFieldNames} from '../actions/updateVisibleFieldNames';
 import {saveViewSettings} from '../utils/saveViewSettings';
+import {updateVisibleFieldNames} from '../views/viewsReducer';
 
 export default function persistVisibleFieldNames({
 	appURL,
